Add ability to change wind model to 'perlin' or 'simple'
Also, moved the instantiation of the WindController to utils/config.py.
Doing this made it so the WindController was only instantiated once and
all references in other files (including tests) was removed.

From now on, all config.wind.speed and config.wind.direction variables
set in config.py will be arrays or Sequence[Sequence[float]].

And set max and min default values in game.py for the wind. Was running
into errors with those variables.